Move and fix tests for DFU BLE transport.
