+++ api/dfu/bleTransport/__tests__/deviceInfoService-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { ErrorCode } = require('../dfuConstants');
+const { ErrorCode } = require('../../dfuConstants');
 const DeviceInfoService = require('../deviceInfoService');
 
 describe('getCharacteristicId', () => {
